+++ src/paco/models/registry.py
@@ -1,0 +1,4 @@
+# Hooks that can be registered by paco.extend
+EXTEND_BASE_MODEL_HOOKS = []
+CW_ALARM_ACTIONS_HOOK = None
+CW_ALARM_HOOKS = []